:white_check_box: :recycle: 화면 이동 테스트 IntegrationTest.kt로 이동
