+++ app/src/androidTest/java/woowacourse/movie/movieDetail/MovieDetailActivityTest.kt
@@ -72,17 +72,4 @@ class MovieDetailActivityTest {
     fun `좌석선택버튼표시_화면에보임`() {
         Espresso.onView(withId(R.id.seat_confirmation_button)).check(matches(isDisplayed()))
     }
-
-    @Test
-    fun `좌석선택버튼클릭_좌석선택표시`() {
-        Espresso.onView(withId(R.id.seat_confirmation_button)).perform(click())
-        Espresso.onView(withId(R.id.seatTable)).check(matches(isDisplayed()))
-    }
-
-    @Test
-    fun `다음화면에서_뒤로가기버튼클릭_영화디테일화면표시`() {
-        Espresso.onView(withId(R.id.seat_confirmation_button)).perform(click())
-        Espresso.pressBack()
-        Espresso.onView(withId(R.id.scroll_view)).check(matches(isDisplayed()))
-    }
 }
